refactor: Update Flask app dependencies
This commit updates the Flask app dependencies in the `requirements.txt` file. The `alembic` package has been added with version 1.13.1, and the `Flask-Cors` and `Flask-Migrate` packages have been added with versions 4.0.1 and 4.0.7, respectively. Additionally, the `Mako` package has been added with version 1.3.5. These updates enhance the functionality and compatibility of the Flask app.