refactor: Add house_section and house_number fields to invoice serializer and model
This commit adds the 'house_section' and 'house_number' fields to the invoice serializer and model in order to capture additional information about the location of the invoice. These fields are required and will be used to improve the accuracy of invoice management. The changes are made in the 'invoice_serializer.py' and 'invoice.py' files.