refactor: Add PrivateRoute component for authenticated routes
This commit adds a new file, PrivateRoute.jsx, which contains the PrivateRoute component. The PrivateRoute component is responsible for rendering the specified component only if the user is authenticated. If the user is not authenticated, they are redirected to the login page. This change improves the security and organization of the routing in the client application.

Affected files:
- client/src/utils/PrivateRoute.jsx